Create a defaultScope for Venue to always exclude the columns createdAt and updatedAt
Change the event price data type to DECIMAL(21, 2)

+++ backend/db/models/event.js
@@ -77,7 +77,7 @@ module.exports = (sequelize, DataTypes) => {
       }
     },
     price: {
-      type: DataTypes.DOUBLE,
+      type: DataTypes.DECIMAL(21, 2),
       allowNull: false,
       validate: {
         isNumeric: true
